fix : Refactor folders structure (app).
