+++ index.html
@@ -14,6 +14,6 @@
 </head>
 <body>
 <div id="root"></div>
-<script type="module" src="/src/app/index.jsx"></script>
+<script type="module" src="/src/index.jsx"></script>
 </body>
 </html>
