+++ src/index.jsx
@@ -3,10 +3,10 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { ThemeWrapper } from '@/modules/Theme';
-import App from './App';
-import './index.css';
+import App from './app/App';
+import './app/index.css';
 
-import store from './store';
+import store from './app/store';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
